requirements.dev.txt
django-extensions==1.5.3
